backend/requirements.txt
# SMART Connect Backend Dependencies
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Data validation and serialization
pydantic==2.5.0
pydantic[email]==2.5.0

# AI Providers
openai==1.3.0
google-generativeai==0.3.0
anthropic==0.7.0

# Genkit integration (if using Firebase Genkit)
# Note: Adjust versions based on your Genkit setup
firebase-admin==6.2.0

# Environment and configuration
python-dotenv==1.0.0

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.0

# Date and time utilities
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.2.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# Production server (optional)
gunicorn==21.2.0

# CORS and middleware
fastapi-cors==0.0.6

# Database migrations
yoyo-migrations==8.2.0

# Background tasks (optional)
celery==5.3.4
redis==5.0.1

# File handling
python-magic==0.4.27

# Excel/CSV processing (for student data import)
pandas==2.1.3
openpyxl==3.1.2

# PDF generation for reports
reportlab==4.0.7

# XML processing
lxml==4.9.3

# Email support (optional)
emails==0.6

# Monitoring and health checks
psutil==5.9.6